+++ src/Sulu/Bundle/SnippetBundle/SuluSnippetBundle.php
@@ -11,8 +11,19 @@
 
 namespace Sulu\Bundle\SnippetBundle;
 
+use Sulu\Bundle\SnippetBundle\DependencyInjection\Compiler\SnippetAreaCompilerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class SuluSnippetBundle extends Bundle
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function build(ContainerBuilder $container)
+    {
+        $container->addCompilerPass(new SnippetAreaCompilerPass());
+
+        parent::build($container);
+    }
 }