+++ src/Sulu/Component/Content/Metadata/StructureMetadata.php
@@ -57,6 +57,11 @@ class StructureMetadata extends ItemMetadata
     public $internal;
 
     /**
+     * @var array
+     */
+    public $areas;
+
+    /**
      * Return a model property.
      *
      * @see StructureMetadata::getProperties()
@@ -211,5 +216,15 @@ class StructureMetadata extends ItemMetadata
     public function isInternal()
     {
         return $this->internal;
+    }
+
+    /**
+     * Return the structure areas.
+     *
+     * @return array
+     */
+    public function getAreas()
+    {
+        return $this->areas;
     }
 }