+++ src/Sulu/Component/Content/Tests/Unit/Metadata/Loader/XmlLegacyLoaderTest.php
@@ -19,7 +19,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
 {
     public function templateDataProvider()
     {
-        return [['page'], ['home'], ['article']];
+        return [['page'], ['home'], ['article'], ['snippet']];
     }
 
     /**
@@ -45,6 +45,20 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
                 'title' => [
                     'de' => 'Das ist das Template 1',
                     'en' => 'That´s the template 1',
+                ],
+            ],
+            'areas' => [
+                [
+                    'key' => 'test',
+                    'title' => [
+                        'de' => 'Test DE',
+                    ],
+                ],
+                [
+                    'key' => 'hotel',
+                    'title' => [
+                        'en' => 'Hotel EN',
+                    ],
                 ],
             ],
             'properties' => [
@@ -266,6 +280,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'page.html.twig',
             'controller' => 'SuluContentBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 2400],
+            'areas' => [],
             'properties' => [
                 'title_section' => [
                     'name' => 'title_section',
@@ -339,6 +354,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'properties' => [],
             'tags' => [],
             'meta' => [],
+            'areas' => [],
         ];
 
         $this->setExpectedException(
@@ -364,6 +380,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'ClientWebsiteBundle:Website:complex.html.twig',
             'controller' => 'SuluWebsiteBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 4800],
+            'areas' => [],
             'properties' => [
                 'title' => [
                     'name' => 'title',
@@ -655,6 +672,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'ClientWebsiteBundle:Website:complex.html.twig',
             'controller' => 'SuluWebsiteBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 4800],
+            'areas' => [],
             'properties' => [
                 'title' => [
                     'name' => 'title',
@@ -873,6 +891,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'page.html.twig',
             'controller' => 'SuluContentBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 2400],
+            'areas' => [],
             'properties' => [
                 'title' => [
                     'name' => 'title',
@@ -1082,6 +1101,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'page.html.twig',
             'controller' => 'SuluContentBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 2400],
+            'areas' => [],
             'properties' => [
                 'title' => [
                     'name' => 'title',
@@ -1148,6 +1168,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'view' => 'page.html.twig',
             'controller' => 'SuluContentBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 2400],
+            'areas' => [],
             'properties' => [
                 'title' => [
                     'name' => 'title',
@@ -1335,6 +1356,7 @@ class XmlLegacyLoaderTest extends \PHPUnit_Framework_TestCase
             'controller' => 'SuluContentBundle:Default:index',
             'cacheLifetime' => ['type' => CacheLifetimeResolverInterface::TYPE_SECONDS, 'value' => 2400],
             'tags' => [],
+            'areas' => [],
             'meta' => [
                 'title' => [
                     'de' => 'Das ist das Template 1',